Exercises 4.15.-4.22

+++ part4/controller/blogs.js
@@ -1,17 +1,37 @@
-const express = require('express')
-const blogRouter = express.Router()
+const blogRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
+const config = require('../utils/config')
 const Blog = require('../model/blogs')
-const { request } = require('express')
+const User = require('../model/users')
+
+const getTokenFrom = (request, response, next) => {
+    const authorization = request.get('authorization')
+    if (authorization && authorization.toLowerCase().startsWith('bearer')) {
+        request.token = authorization.substring(7)
+    }
+    next()
+}
+
+blogRouter.use(getTokenFrom)
 
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({})
+        .populate('user', { username: 1, name: 1, id: 1 })
     response.json(blogs)
 })
 
 blogRouter.post('/', async (request, response) => {
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, config.SECRET)
 
-    if (request.body.title && request.body.url) {
-        const blog = new Blog(request.body)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    if (body.title && body.url) {
+        const blog = new Blog({ ...body, user: user._id })
         const result = await blog.save()
         response.status(201).json(result)
     } else {
@@ -20,8 +40,23 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(200).send()
+    const decodedToken = jwt.verify(request.token, config.SECRET)
+
+    if (!(request.token || !decodedToken)) {
+        return response.status(400).json({ error: 'missing or wrong token' })
+    }
+
+    const blogToEliminate = await Blog.findById(request.params.id)
+    if(!blogToEliminate){
+        return response.status(404).json({ error: 'blog no found' })
+    }
+
+    if (decodedToken.id.toString() === blogToEliminate.user.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(200).send()
+    } else {
+        response.status(401).json({ error: 'does not have credential to eliminate' })
+    }
 })
 
 blogRouter.put('/:id', async (request, response) => {
